Add collection module and refactor shared capture logic into utils
